+++ docs/changelog.rst
@@ -6,6 +6,7 @@ v0.8.0
 
 * Dropped support for Python 3.6
 * Made ``Group.id`` auto-increment
+* Removed CSRF token
 
 v0.7.4
 ^^^^^^